requirements.txt
setuptools~=50.3.0
netcdf4~=1.5.3
numpy~=1.19.2
xarray~=0.16.1
ecmwf-api-client